% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todolist.R
\docType{class}
\name{TodoList-class}
\alias{TodoList}
\alias{TodoList-class}
\title{Reference Class to represent todo list}
\description{
Reference Class to represent todo list
}
\section{Fields}{

\describe{
\item{\code{items}}{List of TodoItems}

\item{\code{nitems}}{Number of items in list}
}}
\section{Methods}{

\describe{
\item{\code{add(text)}}{Generate a new item to the todo list with the given text}

\item{\code{add_item(newItem)}}{Add a TodoItem object to the todo list}
}}

