% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_todo.R
\name{use_todo}
\alias{use_todo}
\title{Get going with TodoList package for your project.}
\usage{
use_todo(name = "todo", file = "./todo.csv")
}
\description{
Get going with TodoList package for your project.
}

